AABB overlap/basic collision checking implemented Changed AABB class to be centerpoint-based Added enum flag to AABB accessors for physics/drawing coord distinction

+++ Egine/StdAfx.h
@@ -67,8 +67,8 @@ inline void SafeRelease( Interface **ppInterfaceToRelease )
 
 typedef struct
 {
-	UINT x;
-	UINT y;
+	double x;
+	double y;
 } PhysPoint;
 
 enum Shape
